✨ [Feat] 기프티콘 취소 시 환불 로직

+++ BE/src/main/java/com/example/ddo_pay/pay/dto/response/GetBalanceResponse.java
@@ -14,9 +14,9 @@ public class GetBalanceResponse {
 
     @JsonIgnore
     private Long userId;
-    private int payBalance;
+    private float payBalance;
 
-    public GetBalanceResponse(int balance) {
+    public GetBalanceResponse(float balance) {
         this.payBalance = balance;
     }
 }
